Data-adaptive test statistics for high-dimensional multiple testing

Performs targeted minimum loss-based estimation (TMLE )of a marginal additive treatment effect of a binary point treatment on an outcome. The data-adaptive algorithm is used to perform variable reduction to avoid the disadvantages associated with multiple testing.

data_adapt_multi_test(Y, A, W = NULL, n.top, n.fold, absolute = FALSE,
  negative = FALSE, parallel = FALSE)

Arguments

Y

continuous or binary outcome variable

A

binary treatment indicator: 1 = treatment, 0 = control

W

matrix containing baseline covariates

n.top

integer value for the number of candidate covariates to generate using the data-adaptive estimation algorithm

n.fold

integer number of folds to be used for cross-validation

absolute

boolean: TRUE = test for absolute effect size. This FALSE = test for directional effect. This overrides argument negative.

negative

boolean: TRUE = test for negative effect size, FALSE = test for positive effect size

parallel

boolean: TRUE = use multiple cores via foreach, FALSE = single-core processing.